refactor(monty): split the runtime into a package, name the OS door for what it is
monty.py was 680 lines holding three unrelated things: the mount view,
the OSAccess subclass monty calls into, and the runtime that drives the
worker pool. They are now modules, matching the wasm tier next door.

    monty/binding.py    the optional pydantic-monty import, in one place
    monty/constants.py  the strings that were inline
    monty/vfs.py        MontyVFS
    monty/osaccess.py   MirageOSAccess, was _MirageOS
    monty/runtime.py    MontyRuntime and the worker helpers

_MirageOS said nothing about what it is. The class is monty's door in
the interception taxonomy: the binding hands the interpreter a host OS
object and calls its methods, so mirage subclasses that object rather
than hooking a syscall layer. MirageOSAccess names that, and the class
is public now because a private name behind a package boundary was
already a contradiction.

INCOMPLETE_MARKERS is the one constant worth naming on its own: the
console tells "keep typing" from "this is broken" by matching monty's
own traceback wording, so a version bump that rephrases either line
turns every continuation into an error at the prompt. Inline it read
like a detail; named, it reads like the coupling it is.

READONLY_HINT moves to wasm/constants.py for the same reason.
