+++ python/mirage/runtime/wasm/constants.py
@@ -12,20 +12,7 @@
 # limitations under the License.
 # ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
-from mirage.runtime.wasm.build import BuildDir
-from mirage.runtime.wasm.config import WasmFsConfig
-from mirage.runtime.wasm.host import WasiFs, install_wasi_fs
-from mirage.runtime.wasm.runtime import WasmRuntime, epoch_engine
-from mirage.runtime.wasm.types import GuestStat
-from mirage.runtime.wasm.vfs import WasmVFS
-
-__all__ = [
-    "BuildDir",
-    "GuestStat",
-    "WasiFs",
-    "WasmFsConfig",
-    "WasmVFS",
-    "WasmRuntime",
-    "epoch_engine",
-    "install_wasi_fs",
-]
+# What a guest is told when it tries to write to the interpreter's own
+# build. Not a permission model: editing the build would change what the
+# next run executes, so the mounts are the only writable half.
+READONLY_HINT = "interpreter build directory is read-only"
